Make `Collider` optional, allowing usage without Parry or Nalgebra
