+++ crates/bevy_xpbd_3d/Cargo.toml
@@ -12,10 +12,18 @@ keywords = ["gamedev", "physics", "simulation", "xpbd", "bevy"]
 categories = ["game-development", "science", "simulation"]
 
 [features]
-default = ["3d", "f32", "async-collider", "debug-plugin", "parallel"]
+default = [
+    "3d",
+    "f32",
+    "parry-f32",
+    "async-collider",
+    "debug-plugin",
+    "parallel",
+]
 3d = []
-f32 = ["dep:parry3d"]
-f64 = ["dep:parry3d-f64"]
+f32 = []
+f64 = []
+
 debug-plugin = ["bevy/bevy_gizmos", "bevy/bevy_render"]
 simd = ["parry3d?/simd-stable", "parry3d-f64?/simd-stable"]
 parallel = ["parry3d?/parallel", "parry3d-f64?/parallel"]
@@ -24,6 +32,13 @@ enhanced-determinism = [
     "parry3d-f64?/enhanced-determinism",
     "glam/libm",
 ]
+
+default-collider = ["dep:nalgebra"]
+# We unfortunately can't reuse the f32 and f64 features for this,
+# because Parry uses separate crates for f32 and f64.
+parry-f32 = ["f32", "dep:parry3d", "default-collider"]
+parry-f64 = ["f64", "dep:parry3d-f64", "default-collider"]
+
 collider-from-mesh = ["bevy/bevy_render"]
 async-collider = ["bevy/bevy_scene", "bevy/bevy_gltf", "collider-from-mesh"]
 serialize = [
@@ -43,7 +58,7 @@ bevy_xpbd_derive = { path = "../bevy_xpbd_derive", version = "0.1" }
 bevy = { version = "0.12", default-features = false }
 parry3d = { version = "0.13", optional = true }
 parry3d-f64 = { version = "0.13", optional = true }
-nalgebra = { version = "0.32", features = ["convert-glam024"] }
+nalgebra = { version = "0.32", features = ["convert-glam024"], optional = true }
 glam = { version = "0.24", features = ["approx"] }
 serde = { version = "1", features = ["derive"], optional = true }
 derive_more = "0.99"
@@ -60,52 +75,57 @@ insta = "1.0"
 
 [[example]]
 name = "dynamic_character_3d"
-required-features = ["3d", "async-collider"]
+required-features = ["3d", "default-collider", "async-collider"]
 
 [[example]]
 name = "kinematic_character_3d"
-required-features = ["3d", "async-collider"]
+required-features = ["3d", "default-collider", "async-collider"]
+
+[[example]]
+name = "cast_ray_predicate"
+required-features = ["3d", "default-collider"]
 
 [[example]]
 name = "chain_3d"
-required-features = ["3d"]
+required-features = ["3d", "default-collider"]
 
 [[example]]
 name = "cubes"
-required-features = ["3d"]
+required-features = ["3d", "default-collider"]
 
 [[example]]
 name = "custom_broad_phase"
-required-features = ["3d"]
+required-features = ["3d", "default-collider"]
 
 [[example]]
 name = "custom_constraint"
-required-features = ["3d"]
+required-features = ["3d", "default-collider"]
 
 [[example]]
 name = "distance_joint_3d"
-required-features = ["3d", "debug-plugin"]
+required-features = ["3d", "default-collider", "debug-plugin"]
 
 [[example]]
 name = "fixed_joint_3d"
-required-features = ["3d"]
+required-features = ["3d", "default-collider"]
 
 [[example]]
 name = "prismatic_joint_3d"
-required-features = ["3d"]
+required-features = ["3d", "default-collider"]
 
 [[example]]
 name = "revolute_joint_3d"
-required-features = ["3d"]
+required-features = ["3d", "default-collider"]
 
 [[example]]
 name = "trimesh_shapes_3d"
-required-features = ["3d"]
+required-features = ["3d", "default-collider"]
 
 [[example]]
 name = "async_colliders"
-required-features = ["3d", "async-collider"]
+required-features = ["3d", "default-collider", "async-collider"]
 
 [[bench]]
 name = "cubes"
+required-features = ["3d", "default-collider"]
 harness = false
