+++ src/plugins/collision/collider/parry.rs
@@ -1,22 +1,10 @@
-use std::fmt;
-
 use crate::{prelude::*, utils::make_isometry};
 #[cfg(all(feature = "3d", feature = "collider-from-mesh"))]
 use bevy::render::mesh::{Indices, VertexAttributeValues};
-#[cfg(all(feature = "3d", feature = "async-collider"))]
-use bevy::utils::HashMap;
-use bevy::{
-    ecs::entity::{EntityMapper, MapEntities},
-    log,
-    prelude::*,
-    utils::HashSet,
-};
+use bevy::{log, prelude::*};
 use collision::contact_query::UnsupportedShape;
 use itertools::Either;
-use parry::{
-    bounding_volume::{Aabb, BoundingVolume},
-    shape::{RoundShape, SharedShape, TypedShape},
-};
+use parry::shape::{RoundShape, SharedShape, TypedShape};
 
 /// Parameters controlling the VHACD convex decomposition algorithm.
 ///
@@ -25,73 +13,6 @@ pub type VHACDParameters = parry::transformation::vhacd::VHACDParameters;
 
 /// Flags used for the preprocessing of a triangle mesh collider.
 pub type TriMeshFlags = parry::shape::TriMeshFlags;
-
-/// A trait that generalizes over colliders. Implementing this trait
-/// allows colliders to be used with the physics engine.
-pub trait AnyCollider: Component {
-    /// Computes the [Axis-Aligned Bounding Box](ColliderAabb) of the collider
-    /// with the given position and rotation.
-    #[cfg_attr(
-        feature = "2d",
-        doc = "\n\nThe rotation is counterclockwise and in radians."
-    )]
-    fn aabb(&self, position: Vector, rotation: impl Into<Rotation>) -> ColliderAabb;
-
-    /// Computes the swept [Axis-Aligned Bounding Box](ColliderAabb) of the collider.
-    /// This corresponds to the space the shape would occupy if it moved from the given
-    /// start position to the given end position.
-    #[cfg_attr(
-        feature = "2d",
-        doc = "\n\nThe rotation is counterclockwise and in radians."
-    )]
-    fn swept_aabb(
-        &self,
-        start_position: Vector,
-        start_rotation: impl Into<Rotation>,
-        end_position: Vector,
-        end_rotation: impl Into<Rotation>,
-    ) -> ColliderAabb {
-        self.aabb(start_position, start_rotation)
-            .merged(self.aabb(end_position, end_rotation))
-    }
-
-    /// Computes the collider's mass properties based on its shape and a given density.
-    fn mass_properties(&self, density: Scalar) -> ColliderMassProperties;
-
-    /// Computes all [`ContactManifold`]s between two colliders.
-    ///
-    /// Returns an empty vector if the colliders are separated by a distance greater than `prediction_distance`
-    /// or if the given shapes are invalid.
-    fn contact_manifolds(
-        &self,
-        other: &Self,
-        position1: Vector,
-        rotation1: impl Into<Rotation>,
-        position2: Vector,
-        rotation2: impl Into<Rotation>,
-        prediction_distance: Scalar,
-    ) -> Vec<ContactManifold>;
-}
-
-/// A trait for colliders that support scaling.
-pub trait ScalableCollider: AnyCollider {
-    /// Returns the global scaling factor of the collider.
-    fn scale(&self) -> Vector;
-
-    /// Sets the global scaling factor of the collider.
-    ///
-    /// If the scaling factor is not uniform and the resulting scaled shape
-    /// can not be represented exactly, the given `detail` is used for an approximation.
-    fn set_scale(&mut self, scale: Vector, detail: u32);
-
-    /// Scales the collider by the given scaling factor.
-    ///
-    /// If the scaling factor is not uniform and the resulting scaled shape
-    /// can not be represented exactly, the given `detail` is used for an approximation.
-    fn scale_by(&mut self, factor: Vector, detail: u32) {
-        self.set_scale(factor * self.scale(), detail)
-    }
-}
 
 /// A collider used for detecting collisions and generating contacts.
 ///
@@ -247,8 +168,8 @@ impl Default for Collider {
     }
 }
 
-impl fmt::Debug for Collider {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+impl std::fmt::Debug for Collider {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self.shape_scaled().as_typed_shape() {
             TypedShape::Ball(shape) => write!(f, "{:?}", shape),
             TypedShape::Cuboid(shape) => write!(f, "{:?}", shape),
@@ -285,10 +206,13 @@ impl fmt::Debug for Collider {
 
 impl AnyCollider for Collider {
     fn aabb(&self, position: Vector, rotation: impl Into<Rotation>) -> ColliderAabb {
-        ColliderAabb(
-            self.shape_scaled()
-                .compute_aabb(&utils::make_isometry(position, rotation)),
-        )
+        let aabb = self
+            .shape_scaled()
+            .compute_aabb(&utils::make_isometry(position, rotation));
+        ColliderAabb {
+            min: aabb.mins.into(),
+            max: aabb.maxs.into(),
+        }
     }
 
     fn mass_properties(&self, density: Scalar) -> ColliderMassProperties {
@@ -1062,391 +986,5 @@ fn scale_shape(
             Ok(SharedShape::compound(scaled))
         }
         _ => Err(parry::query::Unsupported),
-    }
-}
-
-/// A component that will automatically generate a [`Collider`] based on the entity's `Mesh`.
-/// The type of the generated collider can be specified using [`ComputedCollider`].
-///
-/// ## Example
-///
-/// ```
-/// use bevy::prelude::*;
-/// use bevy_xpbd_3d::prelude::*;
-///
-/// fn setup(mut commands: Commands, mut assets: ResMut<AssetServer>, mut meshes: Assets<Mesh>) {
-///     // Spawn a cube with a convex hull collider generated from the mesh
-///     commands.spawn((
-///         AsyncCollider(ComputedCollider::ConvexHull),
-///         PbrBundle {
-///             mesh: meshes.add(Mesh::from(shape::Cube { size: 1.0 })),
-///             ..default()
-///         },
-///     ));
-/// }
-/// ```
-#[cfg(all(feature = "3d", feature = "async-collider"))]
-#[derive(Component, Clone, Debug, Default, Deref, DerefMut)]
-pub struct AsyncCollider(pub ComputedCollider);
-
-/// A component that will automatically generate colliders for the meshes in a scene
-/// once the scene has been loaded. The type of the generated collider can be specified
-/// using [`ComputedCollider`].
-///
-/// ## Example
-///
-/// ```
-/// use bevy::prelude::*;
-/// use bevy_xpbd_3d::prelude::*;
-///
-/// fn setup(mut commands: Commands, mut assets: ResMut<AssetServer>) {
-///     let scene = assets.load("my_model.gltf#Scene0");
-///
-///     // Spawn the scene and automatically generate triangle mesh colliders
-///     commands.spawn((
-///         SceneBundle { scene: scene.clone(), ..default() },
-///         AsyncSceneCollider::new(Some(ComputedCollider::TriMesh)),
-///     ));
-///
-///     // Specify configuration for specific meshes by name
-///     commands.spawn((
-///         SceneBundle { scene: scene.clone(), ..default() },
-///         AsyncSceneCollider::new(Some(ComputedCollider::TriMesh))
-///             .with_shape_for_name("Tree", ComputedCollider::ConvexHull)
-///             .with_layers_for_name("Tree", CollisionLayers::from_bits(0b0010, 0b1111))
-///             .with_density_for_name("Tree", 2.5),
-///     ));
-///
-///     // Only generate colliders for specific meshes by name
-///     commands.spawn((
-///         SceneBundle { scene: scene.clone(), ..default() },
-///         AsyncSceneCollider::new(None)
-///             .with_shape_for_name("Tree", ComputedCollider::ConvexHull),
-///     ));
-///
-///     // Generate colliders for everything except specific meshes by name
-///     commands.spawn((
-///         SceneBundle { scene, ..default() },
-///         AsyncSceneCollider::new(Some(ComputedCollider::TriMeshWithFlags(
-///             TriMeshFlags::MERGE_DUPLICATE_VERTICES
-///         )))
-///         .without_shape_with_name("Tree"),
-///     ));
-/// }
-/// ```
-#[cfg(all(feature = "3d", feature = "async-collider"))]
-#[derive(Component, Clone, Debug, Default, PartialEq)]
-pub struct AsyncSceneCollider {
-    /// The default collider type used for each mesh that isn't included in [`meshes_by_name`](#structfield.meshes_by_name).
-    /// If `None`, all meshes except the ones in [`meshes_by_name`](#structfield.meshes_by_name) will be skipped.
-    pub default_shape: Option<ComputedCollider>,
-    /// Specifies data like the collider type and [`CollisionLayers`] for meshes by name.
-    /// Entries with a `None` value will be skipped.
-    /// For the meshes not found in this `HashMap`, [`default_shape`](#structfield.default_shape)
-    /// and all collision layers will be used instead.
-    pub meshes_by_name: HashMap<String, Option<AsyncSceneColliderData>>,
-}
-
-#[cfg(all(feature = "3d", feature = "async-collider"))]
-impl AsyncSceneCollider {
-    /// Creates a new [`AsyncSceneCollider`] with the default collider type used for
-    /// meshes set to the given `default_shape`.
-    ///
-    /// If the given collider type is `None`, all meshes except the ones in
-    /// [`meshes_by_name`](#structfield.meshes_by_name) will be skipped.
-    /// You can add named shapes using [`with_shape_for_name`](Self::with_shape_for_name).
-    pub fn new(default_shape: Option<ComputedCollider>) -> Self {
-        Self {
-            default_shape,
-            meshes_by_name: default(),
-        }
-    }
-
-    /// Specifies the collider type used for a mesh with the given `name`.
-    pub fn with_shape_for_name(mut self, name: &str, shape: ComputedCollider) -> Self {
-        if let Some(Some(data)) = self.meshes_by_name.get_mut(name) {
-            data.shape = shape;
-        } else {
-            self.meshes_by_name.insert(
-                name.to_string(),
-                Some(AsyncSceneColliderData { shape, ..default() }),
-            );
-        }
-        self
-    }
-
-    /// Specifies the [`CollisionLayers`] used for a mesh with the given `name`.
-    pub fn with_layers_for_name(mut self, name: &str, layers: CollisionLayers) -> Self {
-        if let Some(Some(data)) = self.meshes_by_name.get_mut(name) {
-            data.layers = layers;
-        } else {
-            self.meshes_by_name.insert(
-                name.to_string(),
-                Some(AsyncSceneColliderData {
-                    layers,
-                    ..default()
-                }),
-            );
-        }
-        self
-    }
-
-    /// Specifies the [`ColliderDensity`] used for a mesh with the given `name`.
-    pub fn with_density_for_name(mut self, name: &str, density: Scalar) -> Self {
-        if let Some(Some(data)) = self.meshes_by_name.get_mut(name) {
-            data.density = density;
-        } else {
-            self.meshes_by_name.insert(
-                name.to_string(),
-                Some(AsyncSceneColliderData {
-                    density,
-                    ..default()
-                }),
-            );
-        }
-        self
-    }
-
-    /// Sets collider for the mesh associated with the given `name` to `None`, skipping
-    /// collider generation for it.
-    pub fn without_shape_with_name(mut self, name: &str) -> Self {
-        self.meshes_by_name.insert(name.to_string(), None);
-        self
-    }
-}
-
-/// Configuration for a specific collider generated from a scene using [`AsyncSceneCollider`].
-#[cfg(all(feature = "3d", feature = "async-collider"))]
-#[derive(Clone, Debug, PartialEq)]
-pub struct AsyncSceneColliderData {
-    /// The type of collider generated for the mesh.
-    pub shape: ComputedCollider,
-    /// The [`CollisionLayers`] used for this collider.
-    pub layers: CollisionLayers,
-    /// The [`ColliderDensity`] used for this collider.
-    pub density: Scalar,
-}
-
-#[cfg(all(feature = "3d", feature = "async-collider"))]
-impl Default for AsyncSceneColliderData {
-    fn default() -> Self {
-        Self {
-            shape: ComputedCollider::TriMesh,
-            layers: CollisionLayers::default(),
-            density: 1.0,
-        }
-    }
-}
-
-/// Determines how a [`Collider`] is generated from a `Mesh`.
-///
-/// Colliders can be created from meshes with the following components and methods:
-///
-/// - [`AsyncCollider`] (requires `async-collider` features)
-/// - [`AsyncSceneCollider`] (requires `async-collider` features)
-/// - [`Collider::trimesh_from_mesh`]
-/// - [`Collider::convex_hull_from_mesh`]
-/// - [`Collider::convex_decomposition_from_mesh`]
-#[cfg(all(feature = "3d", feature = "collider-from-mesh"))]
-#[derive(Component, Clone, Debug, Default, PartialEq)]
-pub enum ComputedCollider {
-    /// A triangle mesh.
-    #[default]
-    TriMesh,
-    /// A triangle mesh with a custom configuration.
-    TriMeshWithFlags(TriMeshFlags),
-    /// A convex hull.
-    ConvexHull,
-    /// A compound shape obtained from a decomposition into convex parts using the specified
-    /// [`VHACDParameters`].
-    ConvexDecomposition(VHACDParameters),
-}
-
-/// A component that stores the `Entity` ID of the [`RigidBody`] that a [`Collider`] is attached to.
-///
-/// If the collider is a child of a rigid body, this points to the body's `Entity` ID.
-/// If the [`Collider`] component is instead on the same entity as the [`RigidBody`] component,
-/// this points to the collider's own `Entity` ID.
-///
-/// This component is added and updated automatically based on entity hierarchies and should not
-/// be modified directly.
-///
-/// ## Example
-///
-/// ```
-/// use bevy::prelude::*;
-#[cfg_attr(feature = "2d", doc = "use bevy_xpbd_2d::prelude::*;")]
-#[cfg_attr(feature = "3d", doc = "use bevy_xpbd_3d::prelude::*;")]
-///
-/// fn setup(mut commands: Commands) {
-///     // Spawn a rigid body with one collider on the same entity and two as children.
-///     // Each entity will have a ColliderParent component that has the same rigid body entity.
-///     commands
-///         .spawn((RigidBody::Dynamic, Collider::ball(0.5)))
-///         .with_children(|children| {
-///             children.spawn((Collider::ball(0.5), Transform::from_xyz(2.0, 0.0, 0.0)));
-///             children.spawn((Collider::ball(0.5), Transform::from_xyz(-2.0, 0.0, 0.0)));
-///         });
-/// }
-/// ```
-#[derive(Reflect, Clone, Copy, Component, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-pub struct ColliderParent(pub(crate) Entity);
-
-impl ColliderParent {
-    /// Gets the `Entity` ID of the [`RigidBody`] that this [`Collider`] is attached to.
-    pub const fn get(&self) -> Entity {
-        self.0
-    }
-}
-
-impl MapEntities for ColliderParent {
-    fn map_entities(&mut self, entity_mapper: &mut EntityMapper) {
-        self.0 = entity_mapper.get_or_reserve(self.0)
-    }
-}
-
-/// The transform of a collider relative to the rigid body it's attached to.
-/// This is in the local space of the body, not the collider itself.
-///
-/// This is used for computing things like contact positions and a body's center of mass
-/// without having to traverse deeply nested hierarchies. It's updated automatically,
-/// so you shouldn't modify it manually.
-#[derive(Reflect, Clone, Copy, Component, Debug, PartialEq)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-pub struct ColliderTransform {
-    /// The translation of a collider in a rigid body's frame of reference.
-    pub translation: Vector,
-    /// The rotation of a collider in a rigid body's frame of reference.
-    pub rotation: Rotation,
-    /// The global scale of a collider. Equivalent to the `GlobalTransform` scale.
-    pub scale: Vector,
-}
-
-impl ColliderTransform {
-    /// Transforms a given point by applying the translation, rotation and scale of
-    /// this [`ColliderTransform`].
-    pub fn transform_point(&self, mut point: Vector) -> Vector {
-        point *= self.scale;
-        point = self.rotation.rotate(point);
-        point += self.translation;
-        point
-    }
-}
-
-impl Default for ColliderTransform {
-    fn default() -> Self {
-        Self {
-            translation: Vector::ZERO,
-            rotation: Rotation::default(),
-            scale: Vector::ONE,
-        }
-    }
-}
-
-impl From<Transform> for ColliderTransform {
-    fn from(value: Transform) -> Self {
-        Self {
-            #[cfg(feature = "2d")]
-            translation: value.translation.truncate().adjust_precision(),
-            #[cfg(feature = "3d")]
-            translation: value.translation.adjust_precision(),
-            rotation: Rotation::from(value.rotation.adjust_precision()),
-            #[cfg(feature = "2d")]
-            scale: value.scale.truncate().adjust_precision(),
-            #[cfg(feature = "3d")]
-            scale: value.scale.adjust_precision(),
-        }
-    }
-}
-
-/// A component that marks a [`Collider`] as a sensor, also known as a trigger.
-///
-/// Sensor colliders send [collision events](ContactReportingPlugin#collision-events) and register intersections,
-/// but allow other bodies to pass through them. This is often used to detect when something enters
-/// or leaves an area or is intersecting some shape.
-///
-/// ## Example
-///
-/// ```
-/// use bevy::prelude::*;
-#[cfg_attr(feature = "2d", doc = "use bevy_xpbd_2d::prelude::*;")]
-#[cfg_attr(feature = "3d", doc = "use bevy_xpbd_3d::prelude::*;")]
-///
-/// fn setup(mut commands: Commands) {
-///     // Spawn a static body with a sensor collider.
-///     // Other bodies will pass through, but it will still send collision events.
-///     commands.spawn((RigidBody::Static, Collider::ball(0.5), Sensor));
-/// }
-/// ```
-#[doc(alias = "Trigger")]
-#[derive(Reflect, Clone, Component, Debug, Default, PartialEq, Eq)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-#[reflect(Component)]
-pub struct Sensor;
-
-/// The Axis-Aligned Bounding Box of a [collider](Collider).
-#[derive(Clone, Copy, Component, Debug, Deref, DerefMut, PartialEq)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-pub struct ColliderAabb(pub Aabb);
-
-impl ColliderAabb {
-    /// Creates a new [`ColliderAabb`] from the given `center` and `half_size`.
-    pub fn new(center: Vector, half_size: Vector) -> Self {
-        Self(Aabb::from_half_extents(center.into(), half_size.into()))
-    }
-
-    /// Creates a new [`ColliderAabb`] from a given [`SharedShape`].
-    pub fn from_shape(shape: &SharedShape) -> Self {
-        Self(shape.compute_local_aabb())
-    }
-
-    /// Merges this AABB with another one.
-    pub fn merged(self, other: Self) -> Self {
-        ColliderAabb(self.0.merged(&other.0))
-    }
-}
-
-impl Default for ColliderAabb {
-    fn default() -> Self {
-        ColliderAabb(Aabb::new_invalid())
-    }
-}
-
-/// A component that stores the entities that are colliding with an entity.
-///
-/// This component is automatically added for all entities with a [`Collider`],
-/// but it will only be filled if the [`ContactReportingPlugin`] is enabled (by default, it is).
-///
-/// ## Example
-///
-/// ```
-/// use bevy::prelude::*;
-#[cfg_attr(feature = "2d", doc = "use bevy_xpbd_2d::prelude::*;")]
-#[cfg_attr(feature = "3d", doc = "use bevy_xpbd_3d::prelude::*;")]
-///
-/// fn my_system(query: Query<(Entity, &CollidingEntities)>) {
-///     for (entity, colliding_entities) in &query {
-///         println!(
-///             "{:?} is colliding with the following entities: {:?}",
-///             entity,
-///             colliding_entities
-///         );
-///     }
-/// }
-/// ```
-#[derive(Reflect, Clone, Component, Debug, Default, Deref, DerefMut, PartialEq, Eq)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-#[reflect(Component)]
-pub struct CollidingEntities(pub HashSet<Entity>);
-
-impl MapEntities for CollidingEntities {
-    fn map_entities(&mut self, entity_mapper: &mut EntityMapper) {
-        self.0 = self
-            .0
-            .clone()
-            .into_iter()
-            .map(|e| entity_mapper.get_or_reserve(e))
-            .collect()
     }
 }
